refactor create response method and init execute handler
